Display a diffuse-surfaced ball
This encompasses chapters 1-8 of ray tracing in one weekend, translated into Rust.

Note that I am using sdl2 (https://github.com/Rust-SDL2/rust-sdl2) to display a window, draw pixels and display the result, rather than output the image data to stdout as described in the book. This was mostly just because I wanted to try out creating and drawing to a window in Rust. SDL2 requires some set up in your local environment before you can run this code. I present the canvas regularly during rendering in lieu of a progress bar.

The hit method in the book returned a boolean to indicate whether a hit occurred, and took in a pointer to a hit record to provide the information about the hit if there was one. In Rust, we can instead make use of an enum. We can return either a Hit or a Miss, and include the information about the hit in the Hit enum value. This ends up making the code for handling the result of this method quite nice.

I have implemented a Vec3 struct with various traits to reflect the Vec3 class defined in the book. There are almost certainly more complete linear algebra crates out there, but this was a fun exercise to play with traits.

I am not sure what performance I should expect from this code. I don't even know how long it takes for the C++ code in the book to produce an image. However, it certainly feels very slow. I am far from a Rust expert, and might have made some egregious performance fumbles in here somewhere.