Change diffuse scattering model
As described in section 8.6 of ray tracing in a weekend. The original method we were using to generate a scattered ray was an incorrect approximation to the ideal Lambertian diffuse. The new method instead takes a uniform scatter direction for all angles in a hemisphere centred on the hit point. This seems to result in more severe shadows.