Fix the window not closing during rendering
As I had the rendering event look set up, clicking the window's close
button would not actually close the program while the scene was being
rendered. The escape key still stops the rendering the first time it is
pressed, but I want the close button on the window to actually close it.