Refactoring so that material doesn't know about where it was hit
It annoyed me that a material had to be aware of its position in space to
scatter a ray. Logically, where the material is hit should not matter. All
a material should need to know to scatter a ray is the angle of incidence of
the ray relative to the surface normal. Update the material trait so that
the scatter function only takes in an incident ray and the surface normal.
This means the material can only return a Vec3 direction rather than a ray,
so the caller has to handle constructing the ray. However, this model for a
material makes more sense to me physically, and should make it easier to
add more materials going forward.

Note: I have not read ahead in the ray tracing book to see whether this is
a good idea for future materials, so if you see this commit quickly
reverted, feel free to laugh at me.