Add a movable and orientable camera
Completing chapter 11 of the book. Update our camera so that we can position
and orient it as we like. Not much of interest from a Rust perspective with
this. Mostly just changing the logic in the constructor function for the
camera and updating where it was called. However, it is quite fun to play
around with the parameters for the camera to see how they affect the
resulting rendering of the scene.