Add depth of field
Completing chapter 12 of the book. The code changes were mostly simple to translate to Rust from the original C. Physically, what we are doing is adding a random offset to the origin of every ray that we generate, within the radius of our camera's lens. The example image uses a very wide aperture to exaggerate the effect.